Created a component for search.
